requirements.txt
requests_html == 2.19.1
json == 2.0.9
matplotlib.pyplot == 2.2.2